[Feat][WRS-1897] Show error toast for failed connector's authorization
