+++ src/MainContent.styles.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const MainContentContainer = styled.div`
+    display: flex;
+    height: calc(100vh - 4rem);
+    width: 100%;
+`;
+
+export const CanvasContainer = styled.div`
+    width: 100%;
+`;
