+++ src/components/connector-authentication/index.ts
@@ -1,2 +1,3 @@
-export * from './useConnectorAuthentication';
 export { default as ConnectorAuthenticationModal } from './ConnectorAuthenticationModal';
+export * from './useConnectorAuthentication';
+export * from './useConnectorAuthenticationResult';
